bot-telegram/requirements.txt
httpx==0.27.0
python-dotenv==1.0.1
feedparser==6.0.11
APScheduler==3.10.4